backend/requirements.txt
anyio==3.6.2 ; python_version >= "3.10" and python_version < "4.0"
asgiref==3.6.0 ; python_version >= "3.10" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.10" and python_version < "4.0"
autobahn==23.1.1 ; python_version >= "3.10" and python_version < "4.0"
automat==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.10" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
channels==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==3.0.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
constantly==15.1.0 ; python_version >= "3.10" and python_version < "4.0"
coreapi==2.3.3 ; python_version >= "3.10" and python_version < "4.0"
coreschema==0.0.4 ; python_version >= "3.10" and python_version < "4.0"
coverage==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
coverage[toml]==6.5.0 ; python_version >= "3.10" and python_version < "4.0"
cryptography==39.0.0 ; python_version >= "3.10" and python_version < "4.0"
daphne==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
defusedxml==0.7.1 ; python_version >= "3.10" and python_version < "4.0"
django-cors-headers==3.13.0 ; python_version >= "3.10" and python_version < "4.0"
django-stubs-ext==0.7.0 ; python_version >= "3.10" and python_version < "4.0"
django-stubs==1.14.0 ; python_version >= "3.10" and python_version < "4.0"
django-templated-mail==1.1.1 ; python_version >= "3.10" and python_version < "4.0"
django==4.1.6 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework-simplejwt==4.8.0 ; python_version >= "3.10" and python_version < "4.0"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "4.0"
djoser==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
exceptiongroup==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.10" and python_version < "4.0"
httptools==0.5.0 ; python_version >= "3.10" and python_version < "4.0"
hyperlink==21.0.0 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
incremental==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "4.0"
itypes==1.2.0 ; python_version >= "3.10" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.10" and python_version < "4.0"
mock==4.0.3 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
mypy==0.991 ; python_version >= "3.10" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.10" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.10" and python_version < "4.0"
pillow==9.4.0 ; python_version >= "3.10" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.6.0 ; python_version >= "3.10" and python_version < "4.0"
pyopenssl==23.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-cov==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
pytest-django==4.5.2 ; python_version >= "3.10" and python_version < "4.0"
pytest==7.2.1 ; python_version >= "3.10" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.10" and python_version < "4.0"
python3-openid==3.2.0 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.10" and python_version < "4"
service-identity==21.1.0 ; python_version >= "3.10" and python_version < "4.0"
setuptools==67.1.0 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-app-django==4.0.0 ; python_version >= "3.10" and python_version < "4.0"
social-auth-core==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
sqlparse==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "4.0"
twisted-iocpsupport==1.0.2 ; python_version >= "3.10" and python_version < "4.0" and platform_system == "Windows"
twisted[tls]==22.10.0 ; python_version >= "3.10" and python_version < "4.0"
txaio==23.1.1 ; python_version >= "3.10" and python_version < "4.0"
types-pytz==2022.7.1.0 ; python_version >= "3.10" and python_version < "4.0"
types-pyyaml==6.0.12.4 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.10" and python_version < "4.0"
tzdata==2022.7 ; python_version >= "3.10" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.10" and python_version < "4"
uvicorn[standard]==0.20.0 ; python_version >= "3.10" and python_version < "4.0"
uvloop==0.17.0 ; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.10" and python_version < "4.0"
watchfiles==0.18.1 ; python_version >= "3.10" and python_version < "4.0"
websockets==10.4 ; python_version >= "3.10" and python_version < "4.0"
zope-interface==5.5.2 ; python_version >= "3.10" and python_version < "4.0"
